v0.031: fix scoring bug --
v0.031: fix scoring bug -- wilds weren't getting rotated when the rest of the board was, so for Y-axis words, it was scoring too much if there was a wild already on the board as part of the new word.  (Found while playing Words With Friends: real game scored 2 points less than Games::Literati::wordswithfriends() because a two-point letter was really a wild, but G::L::wwf didn't notice.)  FIX = Swap row/col during wilds-comparison.